refactor: database-module config route and helpers convention
Review feedback: system.database_config is database-module-wide, so its
API moves out of the backups router to GET/PATCH /api/database/config
with the backup settings nested under a `backup` section — future
database-level settings get sibling keys. Shared schema names follow
(GetDatabaseConfigResponse / UpdateDatabaseConfigRequest wrapping the
backup section). Still self-hosting-only, mounted with the backups
router gate.

The backup schedule helpers move from backup-schedule.helpers.ts into
services/database/helpers.ts, matching the one-helpers-file-per-service
convention (payments/ai/email all do the same); their tests fold into
database-helpers.test.ts.
